# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DiyetTakvim API",
    "description": "Diyetisyen–danışan randevu ve beslenme API’si. Aynı route’lar hem `/api/...` hem kökten (`/auth`, `/appointments`, …) erişilebilir. İnteraktif dokümantasyon: `/api-docs`, ham spec: `/openapi.json`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5050",
      "description": "Lokal geliştirme"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Login yanıtındaki access token. Refresh token httpOnly cookie ile gelir."
      }
    },
...
  },
  "paths": {
...
    "/appointments": {
      "post": {
        "tags": ["Randevu"],
        "summary": "Randevu oluştur (danışan)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["dietitianId", "appointmentDate"],
                "properties": {
                  "dietitianId": { "type": "string" },
                  "appointmentDate": { "type": "string", "format": "date-time" },
                  "note": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Oluşturuldu" } }
      }
    },
    "/appointments/available-slots": {
      "get": {
        "tags": ["Randevu"],
        "summary": "Uygun saatler (danışan, bağlı diyetisyen için)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "date", "in": "query", "required": true, "schema": { "type": "string", "example": "2026-04-02" } }],
        "responses": { "200": { "description": "availableSlots" } }
      }
    },
    "/appointments/my-appointments": {
      "get": {
        "tags": ["Randevu"],
        "summary": "Danışan randevularım",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/appointments/dietitian-appointments": {
      "get": {
        "tags": ["Randevu"],
        "summary": "Diyetisyen tüm randevular",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/appointments/daily": {
      "get": {
        "tags": ["Randevu"],
        "summary": "Günlük randevular (diyetisyen)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "date", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/appointments/monthly": {
      "get": {
        "tags": ["Randevu"],
        "summary": "Aylık özet (diyetisyen)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "year", "in": "query", "required": true, "schema": { "type": "integer" } },
          { "name": "month", "in": "query", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "summary" } }
      }
    },
    "/appointments/{id}": {
      "put": {
        "tags": ["Randevu"],
        "summary": "Randevu güncelle",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "appointmentDate": { "type": "string", "format": "date-time" },
                  "note": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "tags": ["Randevu"],
        "summary": "Randevu iptal",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "İptal" } }
      }
    },
...
  }
}